backend/requirements.txt
# FastAPI and web framework
fastapi==0.121.1
uvicorn[standard]==0.24.0
starlette==0.49.3
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.10.6
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.38
asyncpg==0.30.0
supabase==2.3.0

# Cache
redis==5.0.1

# LangChain and AI
langchain==1.0.5
langchain-core==1.0.4
langchain-openai==1.0.2
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
openai==2.7.2
tiktoken==0.12.0

# HTTP client
httpx==0.28.1

# Utilities
python-dotenv==1.2.1
tenacity==9.0.0
yfinance==0.2.28
pandas==2.2.2

# Testing
pytest==9.0.0
pytest-asyncio==1.3.0
pytest-mock==3.15.1

# Additional dependencies
annotated-doc==0.0.4
langsmith==0.4.42
ormsgpack==1.12.0
regex==2025.11.3
tqdm==4.67.1
xxhash==3.6.0
jiter==0.12.0
